Add the OpenCL
Add the OpenCL.
Now it possible to switch between DirectCompute and OpenCL APIs.

+++ Framework/Utils/Timer.h
@@ -10,7 +10,7 @@ public:
 
 	void startMeasure();
 
-	float getMeasuredTime();
+	double getMeasuredTime();
 
 private:
 	Timer();
